+++ NEWS.md
@@ -7,6 +7,7 @@
   * `mapSpain::esp_get_ccaa_siane()`
   * `mapSpain::esp_get_hypsobath()`
   * `mapSpain::esp_get_rivers()`
+  * `mapSpain::esp_get_hydrobasin()`
 
 # mapSpain 0.1.2